requirements.txt


# --- Core Framework ---
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter>=23.3
drf-spectacular==0.26.2

# --- AI / ML Suite ---
# Core ML Libraries
numpy==1.24.4
pandas==2.1.0
scikit-learn==1.3.0
torch==2.1.0
torchvision==0.16.0
tensorflow==2.15.0
ultralytics==8.0.0
transformers==4.35.0
spacy==3.7.0
nltk==3.8.1

# Image/Data Processing
opencv-contrib-python==4.8.1.78
scikit-image>=0.22.0
Pillow==10.1.0
openpyxl>=3.1.0
xlrd>=2.0.1
xlsxwriter>=3.1.0

# --- Database & Storage ---
psycopg2-binary==2.9.7
redis==4.6.0
django-redis>=5.4.0
elasticsearch-dsl>=8.11.0
django-cacheops>=7.0.0
django-storages>=1.14.0
Boto3>=1.34.0

# --- Async & Real-time ---
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0
gunicorn==21.2.0
uvicorn==0.24.0

# --- Authentication & Security ---
djangorestframework-simplejwt==5.3.1
django-oauth-toolkit>=1.7.0
django-ratelimit>=4.1.0
django-axes>=6.1.0
cryptography>=41.0.0
bcrypt>=4.1.0
PyJWT>=2.8.0

# --- Background Tasks ---
celery>=5.3.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0
kombu>=5.3.0
flower>=2.0.0

# --- Utilities & Tools ---
python-decouple==3.8
dj-database-url>=2.1.0
shortuuid>=1.0.0
python-dateutil==2.8.2
pytz>=2023.3
environs>=10.0.0
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0
marshmallow>=3.20.0
pydantic>=2.5.0
python-magic>=0.4.0
reportlab>=4.0.1
weasyprint>=60.1
geopy>=2.4.0

# --- UI & SEO ---
django-meta>=2.0.0
django-imagekit>=4.1.0
django-extensions>=3.2.0
whitenoise>=6.6.0

# --- System Health & Monitoring ---
django-health-check>=3.18.0
django-migration-linter>=5.2.0
django-security-headers>=0.3.1
sentry-sdk>=1.38.0
django-debug-toolbar>=4.2.0
structlog>=23.2.0
django-silk>=5.0.0

# --- Testing & Quality ---
pytest>=7.4.0
pytest-django>=4.7.0
factory-boy>=3.3.0
coverage>=7.3.0
black>=23.11.0
flake8>=6.1.0
isort>=5.12.0
pre-commit>=3.6.0
